requirements.txt
attrs==21.2.0
cattrs==1.10.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
lxml==4.9.1
bs4==0.0.1
python-dateutil==2.8.2
requests==2.31.0
pandas==1.5.2
bert-for-tf2==0.14.9
sentencepiece==0.1.97
transformers==4.31.0
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
datasets==2.14.1
faiss-cpu==1.7.4
diffusers==0.19.2
accelerate==0.21.0
safetensors==0.3.1
invisible-watermark==0.2.0
openai==0.27.8
gym==0.26.2
gym[classic_control]>=0.26.2
matplotlib==3.8.2
pygame>=2.1.0